# openapi.yaml
openapi: 3.0.3
info:
  title: Arepos Server API
  description: REST API для управления моделями, узлами, связями и компонентами
  version: 0.1.1
  contact:
    name: API Support
servers:
  - url: http://arepos-server.arch.svc.cluster.local:8080
    description: Локальный сервер разработки
  - url: https://api.example.com
    description: Продакшн сервер
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/notations:
    get:
      tags:
        - notations
      summary: Список нотаций
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
        - name: ownerId
          in: query
          description: Фильтр по владельцу
          schema:
            type: string
            format: uuid
        - name: name
          in: query
          description: Фильтр по имени (поиск по подстроке)
          schema:
            type: string
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageNotationResponse'
    post:
      tags:
        - notations
      summary: Создать нотацию
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotationRequest'
      responses:
        '201':
          description: Нотация создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotationResponse'
        '404':
          description: Владелец не найден
        '400':
          description: Неверный запрос
  /api/v1/notations/{id}:
    get:
      tags:
        - notations
      summary: Получить нотацию по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotationResponse'
        '404':
          description: Нотация не найдена
    put:
      tags:
        - notations
      summary: Обновить нотацию
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotationUpdateRequest'
      responses:
        '200':
          description: Нотация обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotationResponse'
        '404':
          description: Нотация или владелец не найдены
    delete:
      tags:
        - notations
      summary: Удалить нотацию
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Нотация удалена
        '404':
          description: Нотация не найдена
  /api/v1/notations/{id}/meta:
    get:
      tags:
        - notations
      summary: Получить метаданные нотации по ID
      description: Возвращает имя, версию и владельца нотации для отображения связанной информации в диаграммах.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotationMetaResponse'
        '403':
          description: Недостаточно прав
        '404':
          description: Нотация не найдена
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access токен. Получите через /api/v1/auth/login или /api/v1/auth/register
  schemas:
...
    NotationMetaResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        ownerId:
          type: string
          format: uuid
        ownerEmail:
          type: string
          format: email

    # Аутентификация
...
    PageInfo:
      type: object
      description: Метаданные пагинации (Spring Data Page)
      properties:
        totalElements:
          type: integer
          format: int64
          description: Общее количество элементов
        totalPages:
          type: integer
          description: Общее количество страниц
        size:
          type: integer
          description: Размер страницы
        number:
          type: integer
          description: Номер текущей страницы (с 0)
        first:
          type: boolean
          description: Первая страница
        last:
          type: boolean
          description: Последняя страница
        numberOfElements:
          type: integer
          description: Количество элементов на текущей странице
        empty:
          type: boolean
          description: Пустая страница
        sort:
          type: object
          nullable: true
          description: Параметры сортировки

    # Пользователи
...
    NotationRequest:
      type: object
      required:
        - name
        - version
      properties:
        name:
          type: string
        version:
          type: string
        ownerId:
          type: string
          format: uuid
          description: Если не указан, берется из JWT токена
        attrs:
          type: string
          nullable: true
    NotationUpdateRequest:
      type: object
      properties:
        name:
          type: string
          nullable: true
        version:
          type: string
          nullable: true
        ownerId:
          type: string
          format: uuid
          nullable: true
        attrs:
          type: string
          nullable: true
    NotationResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        ownerId:
          type: string
          format: uuid
        accessPermission:
          type: string
          enum: [ADMIN, OWNER, EDIT, VIEW]
          nullable: true
        attrs:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время создания
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Дата и время последнего обновления
    PageNotationResponse:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/NotationResponse'

    # Типы узлов
...